Add Admin tool for looking up API keys
When an API key is discovered on the public internet, we want to be able to check if it's still active, and then ideally reach out to the API key owner(s). This tool gives us a single place to search all of our private API key types.

I plan to further expand this tool with the ability to email affected users/orgs about their key being leaked, as well as potentially adding the ability to automatically roll/revoke the key.